Add votes router and handlers
