+++ src/routers/votes.ts
@@ -1,0 +1,12 @@
+import type { PostgresJsDatabase } from 'drizzle-orm/postgres-js';
+import { default as express } from 'express';
+import type * as db from '../db';
+import { isLoggedIn } from '../middleware/isLoggedIn';
+import { saveVotesHandler } from '../handlers/votes';
+
+const router = express.Router();
+
+export function votesRouter({ dbPool }: { dbPool: PostgresJsDatabase<typeof db> }) {
+  router.post('/', isLoggedIn(dbPool), saveVotesHandler(dbPool));
+  return router;
+}
